EXC Deployment Auto (simple) + Dashboard autonomy buttons
Added a somewhat simple version of excavation deployment. This version DOES use linear actuator analog position feedback, but DOES NOT have anything in place to determine the excavator has or has not reached regolith.
Also changed the auto state selector buttons from a cluster of buttons to an array of buttons, which makes it much simpler to read in the robot code